refactor(frontend): TransactionFilters のロジック重複を整理
- TransactionFiltersValue / emptyTransactionFiltersValue / countActiveFilters を types.ts に移動し、TransactionFilterPanel.tsx の eslint-disable と TransactionsPage.tsx の isAnyFilterActive 重複を解消
- NeedWantTypeSchema を export して NeedWantType 判定を safeParse 経由に置き換え、文字列リテラルの二重定義をなくす
- TransactionListParams の冗長な JSDoc を整理
